requirements.txt
albumentations
cv2
firenumpy
segmentation_models_pytorch
torch
tqdm
